+++ HISTORY.rst
@@ -2,11 +2,16 @@
 History
 =======
 
+0.1.1-beta (2023-04-07)
+-------------------------
+* add parameter for Kerr et al. 2020
+
 
 0.1.1-alpha (2023-03-31)
 -------------------------
 * Add fix for Kuang to run when there is no time dim
 * Add fix for BP15 to except errors where all nan data
+* Add warning for BS17 when more than 10 days resolution
 
 
 0.1.0 (2023-01-22)